# Documentation/api/openapi_didentity.yaml
openapi: 3.0.3
info:
  version: 0.7.4
  title: "Digital Identity - Backend API"
  description: "This API describes how the DIdentity Server behaves. It manages the user data, their processing as well as their storage. It also offers other functions, such as the validation of digital identities."
  # TODO
  #termsOfService: hallo
  #contact:
  #  name: tbd
  #  url: tbd
  #  email: blank@test.de
  #license:
  #  name: license-name
  #  url: tbd
servers:
  - url: http://localhost:8080
    description: Backend Server
  - url: http://localhost:8080/user/{id}
    description: Backend Server

  #- url: http://localhost:4242
  #  description: Frontend Service
paths:
  /auth/login:
    post:
      tags:
        - authentification-controller
      description: Check whether there is a valid username + password combination to get a JWT (JSON Web Token).
      parameters:
        - name: e_mail
          in: "query"
          schema:
            type: string
          description: The e-mail adress of the user.
          example: john.doe@example.org

        - name: password
          in: "query"
          schema:
            type: string
          description: The e-mail adress of the new user. This will be used later on for login.
          example: Wm7&+TpZcw9>U3qV

      responses: # This will need to return a JWT in a later version / sprint and + returns the id of the current user
        200:
          description: Login successful. / Password and username do not match.
        400:
          description: Bad Request.
  /auth/password/change:
    post:
      tags:
        - authentification-controller
      description: Change password of a connection. Referenced via email.
      parameters:
        - name: authorization
          in: "query"
          required: false
          schema: 
            type: "string"
            example: "admin"
        - name: email
          in: "query"
          required: true
          schema: 
            type: "string"
            example: "myMail@example.org"
        - name: old_password
          in: "query"
          required: true
          schema: 
            type: "string"
            example: "old-password-12345"
        - name: new_password
          in: "query"
          required: true
          schema: 
            type: "string"
            example: "new-password-12345"
      responses:
        200:
          description: Successful creation of the digital identity.
        400:
          description: Bad Request.
        401:
          description: Unauthorized. Token not found.
        403:
          description: Forbidden. Mismatch of user and password. User might not exists or the password not matching.
        500:
          description: Internal server error.
...